fix: ELF_ONLY alone does not make a record an export-table stub
Codex review, P2, and an incompleteness in my own previous fix. Splitting
fact (a) out of the legacy enum closed only one of the two places that read
header absence out of `Visibility.ELF_ONLY`: `is_export_table_only_record`
still keyed on the enum alone. Both header-AST backends assign `ELF_ONLY` to
a declaration they parsed *out of a header* whose symbol turned up in
`.symtab` rather than the dynamic table, so that record -- which has a real
return type, real parameters and a demangled name -- was classified as a
stub. Reproduced: it was dropped from `in_source_declaration_index`, its
parsed signature would be rejected by `bundle_signature_evidence`, and its
removal could be mislabelled `FUNC_REMOVED_ELF_ONLY`.

The enum stays necessary (it is the only thing naming the producer) but is no
longer sufficient: a stub is `ELF_ONLY` *and* not header-declared. Fact (a) is
precisely the evidence that separates the two producers, which is why it had
to stop being derived from the enum before this could be fixed. The DWARF case
the previous revision regressed is untouched -- its (a) is unknown, but it is
not `ELF_ONLY`, so the first clause excludes it; case97 still passes.

One of my own tests had asserted the misclassification, so it was pinning the
bug; it now asserts the correct reading, with the reason recorded inline.

The test file reached the 1200-line cap for a new test file, so the
export-axis *detector* classes move to `tests/test_export_transition.py`,
giving `abicheck/compare/export_transition.py` the test file it should have
had: those classes drive `checker.compare` and assert findings, where the
siblings that stay assert what a fact means. Test count is conserved exactly
at 484 -- a first attempt at the split silently dropped the 343-case
exhaustive class, which is why the count is now checked rather than assumed.
